Create monitor
Intent behind the monitor is that it logs the traffic coming from traffic.py into our target port, Radar Station 1. This traffic inevitably works its way to the Command Center. The command center than collates Radar Station 1's data with other data from the other radar stations, and then sends it all back out to the focal point, which then routes this data (minus what the radar station first recorded) back to the Radar Station. This is supposed to fill in the gaps for what that radar station is missing. How this is represented is that the monitor logs what it receives along port 5562, the port that focal point uses to communicate to radar station 1 in radar stations.py.