Update Command Center.py
Added a lot. Added 5 ports for the command center to use to connect to the focal point, one for each of the radar stations and carriers. 

Added a filter function, which is sort of the crux of the project, so that's good. Forwards traffic to every station besides that which it originated from.